requirements.txt
black
flake8
ipython
isort
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
scikit-learn
tqdm
typer
-e .
dvc
seaborn
pyyaml
xgboost 